Changes to alignment class
